Refactor routes store
The routes store keeps a tree of route objects, initialized from the
<Routes> component just before it's inserted into the DOM. Instead of
using components for finding active routes when the URL changes, it
uses the tree representation instead. If <Route> component props are
updated, this won't change lookups, which works more like a true Flux
store which requires explicit actions to change data in the store.

Also, added separate getActiveRoutes and getActiveParams methods.