Upgrade to lerna 3 + stop hoisting deps
The hoisting was making it difficult for me to keep track of whether or
not a package had its own dependencies listed correctly in its own
package.json or if it was re-using a dependency specified by some other
package, so I disabled it for now. If we would still like this behavior,
I'd be open to moving all devDependencies into the root package.json so
that they are only installed once. But for now, the extra time it takes
on the install is worth it for me.